requirements.txt


selenium
beautifulsoup4
psycopg2
python-dotenv
